// workspaces/ai-integrations/plugins/catalog-backend-module-kagenti/src/providers/types.ts
/*
 * Copyright Red Hat, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import type { SchedulerServiceTaskScheduleDefinition } from '@backstage/backend-plugin-api';

/**
 * KagentiConfig defines the configuration for the Kagenti provider
 *
 * @public
 */
export type KagentiConfig = {
  id: string;
  baseUrl: string;
  clientId: string;
  clientSecret: string;
  tokenUrl: string;
  namespaces?: string[];
  lifecycleMapping?: Record<string, string>;
  schedule?: SchedulerServiceTaskScheduleDefinition;
};